pyproject.toml
[project]
name = "tick-telegram-bot"
version = "0.1.0"
description = "A simple Telegram todo bot using local JSON storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "logger>=1.4",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=22.5",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tick_telegram_bot"]
package-dir = "src"

[tool.hatch.build.targets.sdist]
include = ["src/tick_telegram_bot"]

[tool.uv]
managed = true
